fix: Handle case when removed issue key does not exist
As both drivers when it's restarted remove their content, it might be a case when a client tries to trigger the delete operation after the backend restart. In such a case it led to a situation when the backend application crashed as the drivers were not able to handle the described situation correctly.